perf: NestJS REST API 튜닝 및 Phase 4 부하테스트 리포트 정리
- N+1 쿼리 제거: getAllActiveChatRooms() 400→5쿼리, getChatRoomMessages() 300→4쿼리
- 배치 로드 도입: findByIds, findByMessageIds, findActiveByChatRoomIds
- TypeORM 커넥션풀 10→50 확장
- DB 인덱스 추가: idx_msg_room_deleted, idx_msg_room_sent, idx_msg_sender, idx_room_active
- 부하테스트 리포트 Phase별 분리 (phase-1~4.md)
- 튜닝 후 결과: HTTP avg 1,165→767ms(-34%), p95 2,262→1,870ms(-17%)

+++ server/chat-server/src/infrastructure/persistence/entity/chat-room-member.entity.ts
@@ -33,6 +33,6 @@ export class ChatRoomMemberEntity {
   @Column({ name: 'last_active_at', type: 'timestamp' })
   lastActiveAt!: Date;
 
-  @Column({ name: 'last_read_message_id', type: 'bigint', nullable: true })
+  @Column({ name: 'last_read_message_id', type: 'int', nullable: true })
   lastReadMessageId: number | null = null;
 }
